move functions to create instances of models to new file with scripts
