+++ StudentPlatform/platformapp/tests/utils_for_testing.py
@@ -2,8 +2,6 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.shortcuts import reverse
 
-from ..models import Group, Tab, Element
-
 User = get_user_model()
 LOGIN_URL = reverse('login_view')
 
@@ -31,48 +29,6 @@ def create_two_users_authenticate_one(test_case: TestCase):
     test_case.logged_user = create_user('logged', 'logged')
     test_case.not_logged_user = create_user('notlogged', 'notlogged')
     test_case.client.login(username='logged', password='logged')
-
-
-def create_group(name, description, user):
-    """Create group for testing purposes"""
-
-    group = Group.objects.create(
-        name=name,
-        description=description,
-        creator=user,
-    )
-    group.save()
-    group.users.add(user)
-    group.save()
-
-    return group
-
-
-def create_tab(name, user, group):
-    """Create Tab for testing purposes"""
-
-    tab = Tab.objects.create(
-        name=name,
-        creator=user,
-        group=group,
-    )
-    tab.save()
-
-    return tab
-
-
-def create_element(name, text, user, tab):
-    """Create element for testing purposes"""
-
-    element = Element.objects.create(
-        name=name,
-        text=text,
-        creator=user,
-        tab=tab,
-    )
-    element.save()
-
-    return element
 
 
 def login_redirect_url(url):
